refactor(configurator): split daemon page modules
